remove project settings too
