+++ client/ayon_core/settings/__init__.py
@@ -1,22 +1,18 @@
-from .constants import (
-    PROJECT_SETTINGS_KEY,
-)
 from .lib import (
-    get_general_environments,
+    get_ayon_settings,
+    get_studio_settings,
     get_system_settings,
     get_project_settings,
+    get_general_environments,
     get_current_project_settings,
 )
-from .ayon_settings import get_ayon_settings
 
 
 __all__ = (
-    "PROJECT_SETTINGS_KEY",
-
-    "get_general_environments",
+    "get_ayon_settings",
+    "get_studio_settings",
     "get_system_settings",
+    "get_general_environments",
     "get_project_settings",
     "get_current_project_settings",
-
-    "get_ayon_settings",
 )
