--- client/ayon_core/settings/defaults/project_anatomy/attributes.json
@@ -1,26 +1,0 @@
-{
-    "fps": 25.0,
-    "frameStart": 1001,
-    "frameEnd": 1001,
-    "clipIn": 1,
-    "clipOut": 1,
-    "handleStart": 0,
-    "handleEnd": 0,
-    "resolutionWidth": 1920,
-    "resolutionHeight": 1080,
-    "pixelAspect": 1.0,
-    "applications": [
-        "maya/2020",
-        "nuke/12-2",
-        "nukex/12-2",
-        "hiero/12-2",
-        "resolve/stable",
-        "houdini/18-5",
-        "blender/2-91",
-        "harmony/20",
-        "photoshop/2021",
-        "aftereffects/2021"
-    ],
-    "tools_env": [],
-    "active": true
-}
