--- client/ayon_core/settings/defaults/project_anatomy/roots.json
@@ -1,7 +1,0 @@
-{
-    "work": {
-        "windows": "C:/projects",
-        "darwin": "/Volumes/path",
-        "linux": "/mnt/share/projects"
-    }
-}
