--- client/ayon_core/settings/defaults/project_anatomy/tasks.json
@@ -1,44 +1,0 @@
-{
-    "Generic": {
-        "short_name": "gener"
-    },
-    "Art": {
-        "short_name": "art"
-    },
-    "Modeling": {
-        "short_name": "mdl"
-    },
-    "Texture": {
-        "short_name": "tex"
-    },
-    "Lookdev": {
-        "short_name": "look"
-    },
-    "Rigging": {
-        "short_name": "rig"
-    },
-    "Edit": {
-        "short_name": "edit"
-    },
-    "Layout": {
-        "short_name": "lay"
-    },
-    "Setdress": {
-        "short_name": "dress"
-    },
-    "Animation": {
-        "short_name": "anim"
-    },
-    "FX": {
-        "short_name": "fx"
-    },
-    "Lighting": {
-        "short_name": "lgt"
-    },
-    "Paint": {
-        "short_name": "paint"
-    },
-    "Compositing": {
-        "short_name": "comp"
-    }
-}
