--- client/ayon_core/settings/defaults/project_anatomy/templates.json
@@ -1,67 +1,0 @@
-{
-    "defaults": {
-        "version_padding": 3,
-        "version": "v{version:0>{@version_padding}}",
-        "frame_padding": 4,
-        "frame": "{frame:0>{@frame_padding}}"
-    },
-    "work": {
-        "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/work/{task[name]}",
-        "file": "{project[code]}_{asset}_{task[name]}_{@version}<_{comment}>.{ext}",
-        "path": "{@folder}/{@file}"
-    },
-    "render": {
-        "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/publish/{family}/{subset}/{@version}",
-        "file": "{project[code]}_{asset}_{subset}_{@version}<_{output}><.{@frame}>.{ext}",
-        "path": "{@folder}/{@file}"
-    },
-    "publish": {
-        "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/publish/{family}/{subset}/{@version}",
-        "file": "{project[code]}_{asset}_{subset}_{@version}<_{output}><.{@frame}><_{udim}>.{ext}",
-        "path": "{@folder}/{@file}",
-        "thumbnail": "{thumbnail_root}/{project[name]}/{_id}_{thumbnail_type}.{ext}"
-    },
-    "hero": {
-        "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/publish/{family}/{subset}/hero",
-        "file": "{project[code]}_{asset}_{subset}_hero<_{output}><.{frame}>.{ext}",
-        "path": "{@folder}/{@file}"
-    },
-    "delivery": {},
-    "unreal": {
-        "folder": "{root[work]}/{project[name]}/unreal/{task[name]}",
-        "file": "{project[code]}_{asset}.{ext}",
-        "path": "{@folder}/{@file}"
-    },
-    "others": {
-        "maya2unreal": {
-            "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/publish/{family}",
-            "file": "{subset}_{@version}<_{output}><.{@frame}>.{ext}",
-            "path": "{@folder}/{@file}"
-        },
-        "online": {
-            "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/publish/{family}/{subset}/{@version}",
-            "file": "{originalBasename}<.{@frame}><_{udim}>.{ext}",
-            "path": "{@folder}/{@file}"
-        },
-        "tycache": {
-            "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/publish/{family}/{subset}/{@version}",
-            "file": "{originalBasename}.{ext}",
-            "path": "{@folder}/{@file}"
-        },
-        "source": {
-            "folder": "{root[work]}/{originalDirname}",
-            "file": "{originalBasename}.{ext}",
-            "path": "{@folder}/{@file}"
-        },
-        "transient": {
-            "folder": "{root[work]}/{project[name]}/{hierarchy}/{asset}/work/{family}/{subset}"
-        },
-        "__dynamic_keys_labels__": {
-            "maya2unreal": "Maya to Unreal",
-            "online": "online",
-            "tycache": "tycache",
-            "source": "source",
-            "transient": "transient"
-        }
-    }
-}
