--- client/ayon_core/settings/defaults/project_settings/aftereffects.json
@@ -1,49 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "create": {
-        "RenderCreator": {
-            "default_variants": [
-                "Main"
-            ],
-            "mark_for_review": true
-        }
-    },
-    "publish": {
-        "CollectReview": {
-            "enabled": true
-        },
-        "ValidateSceneSettings": {
-            "enabled": true,
-            "optional": true,
-            "active": true,
-            "skip_resolution_check": [
-                ".*"
-            ],
-            "skip_timelines_check": [
-                ".*"
-            ]
-        },
-        "ValidateContainers": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        }
-    },
-    "workfile_builder": {
-        "create_first_version": false,
-        "custom_templates": []
-    },
-    "templated_workfile_build": {
-        "profiles": []
-    }
-}
