--- client/ayon_core/settings/defaults/project_settings/applications.json
@@ -1,3 +1,0 @@
-{
-    "only_available": false
-}
