--- client/ayon_core/settings/defaults/project_settings/blender.json
@@ -1,217 +1,0 @@
-{
-    "unit_scale_settings": {
-        "enabled": true,
-        "apply_on_opening": false,
-        "base_file_unit_scale": 0.01
-    },
-    "set_resolution_startup": true,
-    "set_frames_startup": true,
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "RenderSettings": {
-        "default_render_image_folder": "renders/blender",
-        "aov_separator": "underscore",
-        "image_format": "exr",
-        "multilayer_exr": true,
-        "renderer": "CYCLES",
-        "aov_list": ["combined"],
-        "custom_passes": []
-    },
-    "workfile_builder": {
-        "create_first_version": false,
-        "custom_templates": []
-    },
-    "publish": {
-        "ValidateCameraZeroKeyframe": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateFileSaved": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "exclude_families": []
-        },
-        "ValidateRenderCameraIsSet": {
-            "enabled": true,
-            "optional": false,
-            "active": true
-        },
-        "ValidateDeadlinePublish": {
-            "enabled": true,
-            "optional": false,
-            "active": true
-        },
-        "ValidateMeshHasUvs": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateMeshNoNegativeScale": {
-            "enabled": true,
-            "optional": false,
-            "active": true
-        },
-        "ValidateTransformZero": {
-            "enabled": true,
-            "optional": false,
-            "active": true
-        },
-        "ValidateNoColonsInName": {
-            "enabled": true,
-            "optional": false,
-            "active": true
-        },
-        "ValidateInstanceEmpty": {
-            "enabled": true,
-            "optional": false,
-            "active": true
-        },
-        "ExtractBlend": {
-            "enabled": true,
-            "optional": true,
-            "active": true,
-            "families": [
-                "model",
-                "camera",
-                "rig",
-                "action",
-                "layout",
-                "blendScene"
-            ]
-        },
-        "ExtractFBX": {
-            "enabled": true,
-            "optional": true,
-            "active": false
-        },
-        "ExtractModelABC": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ExtractBlendAnimation": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ExtractAnimationFBX": {
-            "enabled": true,
-            "optional": true,
-            "active": false
-        },
-        "ExtractCamera": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ExtractCameraABC": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ExtractLayout": {
-            "enabled": true,
-            "optional": true,
-            "active": false
-        },
-        "ExtractThumbnail": {
-            "enabled": true,
-            "optional": true,
-            "active": true,
-            "presets": {
-                "model": {
-                    "image_settings": {
-                        "file_format": "JPEG",
-                        "color_mode": "RGB",
-                        "quality": 100
-                    },
-                    "display_options": {
-                        "shading": {
-                            "light": "STUDIO",
-                            "studio_light": "Default",
-                            "type": "SOLID",
-                            "color_type": "OBJECT",
-                            "show_xray": false,
-                            "show_shadows": false,
-                            "show_cavity": true
-                        },
-                        "overlay": {
-                            "show_overlays": false
-                        }
-                    }
-                },
-                "rig": {
-                    "image_settings": {
-                        "file_format": "JPEG",
-                        "color_mode": "RGB",
-                        "quality": 100
-                    },
-                    "display_options": {
-                        "shading": {
-                            "light": "STUDIO",
-                            "studio_light": "Default",
-                            "type": "SOLID",
-                            "color_type": "OBJECT",
-                            "show_xray": true,
-                            "show_shadows": false,
-                            "show_cavity": false
-                        },
-                        "overlay": {
-                            "show_overlays": true,
-                            "show_ortho_grid": false,
-                            "show_floor": false,
-                            "show_axis_x": false,
-                            "show_axis_y": false,
-                            "show_axis_z": false,
-                            "show_text": false,
-                            "show_stats": false,
-                            "show_cursor": false,
-                            "show_annotation": false,
-                            "show_extras": false,
-                            "show_relationship_lines": false,
-                            "show_outline_selected": false,
-                            "show_motion_paths": false,
-                            "show_object_origins": false,
-                            "show_bones": true
-                        }
-                    }
-                }
-            }
-        },
-        "ExtractPlayblast": {
-            "enabled": true,
-            "optional": true,
-            "active": true,
-            "presets": {
-                "default": {
-                    "image_settings": {
-                        "file_format": "PNG",
-                        "color_mode": "RGB",
-                        "color_depth": "8",
-                        "compression": 15
-                    },
-                    "display_options": {
-                        "shading": {
-                            "type": "MATERIAL",
-                            "render_pass": "COMBINED"
-                        },
-                        "overlay": {
-                            "show_overlays": false
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
