--- client/ayon_core/settings/defaults/project_settings/celaction.json
@@ -1,27 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "workfile": {
-        "submission_overrides": [
-            "render_chunk",
-            "frame_range",
-            "resolution"
-        ]
-    },
-    "publish": {
-        "CollectRenderPath": {
-            "output_extension": "png",
-            "anatomy_template_key_render_files": "render",
-            "anatomy_template_key_metadata": "render"
-        }
-    }
-}
