--- client/ayon_core/settings/defaults/project_settings/deadline.json
@@ -1,157 +1,0 @@
-{
-    "deadline_servers": [],
-    "publish": {
-        "CollectDefaultDeadlineServer": {
-            "pass_mongo_url": true
-        },
-        "CollectDeadlinePools": {
-            "primary_pool": "",
-            "secondary_pool": ""
-        },
-        "ValidateExpectedFiles": {
-            "enabled": true,
-            "active": true,
-            "allow_user_override": true,
-            "families": [
-                "render"
-            ],
-            "targets": [
-                "deadline"
-            ]
-        },
-        "MayaSubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "tile_assembler_plugin": "DraftTileAssembler",
-            "use_published": true,
-            "import_reference": false,
-            "asset_dependencies": true,
-            "priority": 50,
-            "tile_priority": 50,
-            "group": "none",
-            "limit": [],
-            "jobInfo": {},
-            "pluginInfo": {},
-            "scene_patches": [],
-            "strict_error_checking": true
-        },
-        "MaxSubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "use_published": true,
-            "priority": 50,
-            "chunk_size": 10,
-            "group": "none"
-        },
-        "FusionSubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "priority": 50,
-            "chunk_size": 10,
-            "concurrent_tasks": 1,
-            "group": "",
-            "plugin": "Fusion"
-        },
-        "NukeSubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "priority": 50,
-            "chunk_size": 10,
-            "concurrent_tasks": 1,
-            "group": "",
-            "department": "",
-            "use_gpu": true,
-            "workfile_dependency": true,
-            "use_published_workfile": true,
-            "env_allowed_keys": [],
-            "env_search_replace_values": {},
-            "limit_groups": {}
-        },
-        "HarmonySubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "use_published": true,
-            "priority": 50,
-            "chunk_size": 10000,
-            "group": "",
-            "department": ""
-        },
-        "AfterEffectsSubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "use_published": true,
-            "priority": 50,
-            "chunk_size": 10000,
-            "group": "",
-            "department": "",
-            "multiprocess": true
-        },
-        "CelactionSubmitDeadline": {
-            "enabled": true,
-            "deadline_department": "",
-            "deadline_priority": 50,
-            "deadline_pool": "",
-            "deadline_pool_secondary": "",
-            "deadline_group": "",
-            "deadline_chunk_size": 10,
-            "deadline_job_delay": "00:00:00:00"
-        },
-        "BlenderSubmitDeadline": {
-            "enabled": true,
-            "optional": false,
-            "active": true,
-            "use_published": true,
-            "priority": 50,
-            "chunk_size": 10,
-            "group": "none",
-            "job_delay": "00:00:00:00"
-        },
-        "ProcessSubmittedCacheJobOnFarm": {
-            "enabled": true,
-            "deadline_department": "",
-            "deadline_pool": "",
-            "deadline_group": "",
-            "deadline_chunk_size": 1,
-            "deadline_priority": 50
-        },
-        "ProcessSubmittedJobOnFarm": {
-            "enabled": true,
-            "deadline_department": "",
-            "deadline_pool": "",
-            "deadline_group": "",
-            "deadline_chunk_size": 1,
-            "deadline_priority": 50,
-            "publishing_script": "",
-            "skip_integration_repre_list": [],
-            "aov_filter": {
-                "maya": [
-                    ".*([Bb]eauty).*"
-                ],
-                "blender": [
-                    ".*([Bb]eauty).*"
-                ],
-                "aftereffects": [
-                    ".*"
-                ],
-                "celaction": [
-                    ".*"
-                ],
-                "harmony": [
-                    ".*"
-                ],
-                "max": [
-                    ".*"
-                ],
-                "fusion": [
-                    ".*"
-                ]
-            }
-        }
-    }
-}
