--- client/ayon_core/settings/defaults/project_settings/flame.json
@@ -1,148 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "remapping": {
-            "rules": []
-        },
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        },
-        "project": {
-            "colourPolicy": "ACES 1.1",
-            "frameDepth": "16-bit fp",
-            "fieldDominance": "PROGRESSIVE"
-        },
-        "profilesMapping": {
-            "inputs": [
-                {
-                    "flameName": "ACEScg",
-                    "ocioName": "ACES - ACEScg"
-                },
-                {
-                    "flameName": "Rec.709 video",
-                    "ocioName": "Output - Rec.709"
-                }
-            ]
-        }
-    },
-    "create": {
-        "CreateShotClip": {
-            "hierarchy": "{folder}/{sequence}",
-            "useShotName": true,
-            "clipRename": false,
-            "clipName": "{sequence}{shot}",
-            "segmentIndex": true,
-            "countFrom": 10,
-            "countSteps": 10,
-            "folder": "shots",
-            "episode": "ep01",
-            "sequence": "a",
-            "track": "{_track_}",
-            "shot": "####",
-            "vSyncOn": false,
-            "workfileFrameStart": 1001,
-            "handleStart": 5,
-            "handleEnd": 5,
-            "includeHandles": false,
-            "retimedHandles": true,
-            "retimedFramerange": true
-        }
-    },
-    "publish": {
-        "CollectTimelineInstances": {
-            "xml_preset_attrs_from_comments": [
-                {
-                    "name": "width",
-                    "type": "number"
-                },
-                {
-                    "name": "height",
-                    "type": "number"
-                },
-                {
-                    "name": "pixelRatio",
-                    "type": "float"
-                },
-                {
-                    "name": "resizeType",
-                    "type": "string"
-                },
-                {
-                    "name": "resizeFilter",
-                    "type": "string"
-                }
-            ],
-            "add_tasks": [
-                {
-                    "name": "compositing",
-                    "type": "Compositing",
-                    "create_batch_group": true
-                }
-            ]
-        },
-        "ExtractSubsetResources": {
-            "keep_original_representation": false,
-            "export_presets_mapping": {
-                "exr16fpdwaa": {
-                    "active": true,
-                    "export_type": "File Sequence",
-                    "ext": "exr",
-                    "xml_preset_file": "OpenEXR (16-bit fp DWAA).xml",
-                    "colorspace_out": "ACES - ACEScg",
-                    "xml_preset_dir": "",
-                    "parsed_comment_attrs": true,
-                    "representation_add_range": true,
-                    "representation_tags": [],
-                    "load_to_batch_group": true,
-                    "batch_group_loader_name": "LoadClipBatch",
-                    "filter_path_regex": ".*"
-                }
-            }
-        },
-        "IntegrateBatchGroup": {
-            "enabled": false
-        }
-    },
-    "load": {
-        "LoadClip": {
-            "enabled": true,
-            "families": [
-                "render2d",
-                "source",
-                "plate",
-                "render",
-                "review"
-            ],
-            "reel_group_name": "OpenPype_Reels",
-            "reel_name": "Loaded",
-            "clip_name_template": "{asset}_{subset}<_{output}>",
-            "layer_rename_template": "{asset}_{subset}<_{output}>",
-            "layer_rename_patterns": [
-                "rgb",
-                "rgba"
-            ]
-        },
-        "LoadClipBatch": {
-            "enabled": true,
-            "families": [
-                "render2d",
-                "source",
-                "plate",
-                "render",
-                "review"
-            ],
-            "reel_name": "OP_LoadedReel",
-            "clip_name_template": "{batch}_{asset}_{subset}<_{output}>",
-            "layer_rename_template": "{asset}_{subset}<_{output}>",
-            "layer_rename_patterns": [
-                "rgb",
-                "rgba"
-            ]
-        }
-    }
-}
