--- client/ayon_core/settings/defaults/project_settings/fusion.json
@@ -1,58 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "copy_fusion_settings": {
-        "copy_path": "~/.openpype/hosts/fusion/profiles",
-        "copy_status": false,
-        "force_sync": false
-    },
-    "hooks": {
-        "InstallPySideToFusion": {
-            "enabled": true
-        }
-    },
-    "create": {
-        "CreateSaver": {
-            "temp_rendering_path_template": "{workdir}/renders/fusion/{subset}/{subset}.{frame}.{ext}",
-            "default_variants": [
-                "Main",
-                "Mask"
-            ],
-            "instance_attributes": [
-                "reviewable",
-                "farm_rendering"
-            ],
-            "image_format": "exr",
-            "default_frame_range_option": "asset_db"
-        },
-        "CreateImageSaver": {
-            "temp_rendering_path_template": "{workdir}/renders/fusion/{subset}/{subset}.{ext}",
-            "default_variants": [
-                "Main",
-                "Mask"
-            ],
-            "instance_attributes": [
-                "reviewable",
-                "farm_rendering"
-            ],
-            "image_format": "exr",
-            "default_frame": 0
-        }
-    },
-    "publish": {
-        "ValidateSaverResolution": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        }
-    }
-}
