--- client/ayon_core/settings/defaults/project_settings/harmony.json
@@ -1,38 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "publish": {
-        "CollectPalettes": {
-            "allowed_tasks": [
-                ".*"
-            ]
-        },
-        "ValidateAudio": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateContainers": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateSceneSettings": {
-            "enabled": true,
-            "optional": true,
-            "active": true,
-            "frame_check_filter": [],
-            "skip_resolution_check": [],
-            "skip_timelines_check": []
-        }
-    }
-}
