--- client/ayon_core/settings/defaults/project_settings/hiero.json
@@ -1,87 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        },
-        "workfile": {
-            "ocioConfigName": "nuke-default",
-            "workingSpace": "linear",
-            "sixteenBitLut": "sRGB",
-            "eightBitLut": "sRGB",
-            "floatLut": "linear",
-            "logLut": "Cineon",
-            "viewerLut": "sRGB",
-            "thumbnailLut": "sRGB",
-            "monitorOutLut": "sRGB"
-        },
-        "regexInputs": {
-            "inputs": [
-                {
-                    "regex": "[^-a-zA-Z0-9](plateRef).*(?=mp4)",
-                    "colorspace": "sRGB"
-                }
-            ]
-        }
-    },
-    "create": {
-        "CreateShotClip": {
-            "hierarchy": "{folder}/{sequence}",
-            "clipRename": true,
-            "clipName": "{track}{sequence}{shot}",
-            "countFrom": 10,
-            "countSteps": 10,
-            "folder": "shots",
-            "episode": "ep01",
-            "sequence": "sq01",
-            "track": "{_track_}",
-            "shot": "sh###",
-            "vSyncOn": false,
-            "workfileFrameStart": 1001,
-            "handleStart": 10,
-            "handleEnd": 10
-        }
-    },
-    "load": {
-        "LoadClip": {
-            "enabled": true,
-            "families": [
-                "render2d",
-                "source",
-                "plate",
-                "render",
-                "review"
-            ],
-            "clip_name_template": "{asset}_{subset}_{representation}"
-        }
-    },
-    "publish": {
-        "CollectInstanceVersion": {
-            "enabled": false
-        },
-        "ExtractReviewCutUpVideo": {
-            "enabled": true,
-            "tags_addition": [
-                "review"
-            ]
-        }
-    },
-    "filters": {},
-    "scriptsmenu": {
-        "name": "OpenPype Tools",
-        "definition": [
-            {
-                "type": "action",
-                "sourcetype": "python",
-                "title": "OpenPype Docs",
-                "command": "import webbrowser;webbrowser.open(url='https://openpype.io/docs/artist_hosts_hiero')",
-                "tooltip": "Open the OpenPype Hiero user doc page"
-            }
-        ]
-    }
-}
