--- client/ayon_core/settings/defaults/project_settings/houdini.json
@@ -1,192 +1,0 @@
-{
-    "general": {
-        "add_self_publish_button": false,
-        "update_houdini_var_context": {
-            "enabled": true,
-            "houdini_vars":[
-                {
-                    "var": "JOB",
-                    "value": "{root[work]}/{project[name]}/{hierarchy}/{asset}/work/{task[name]}",
-                    "is_directory": true
-                }
-            ]
-        }
-    },
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "shelves": [],
-    "create": {
-        "CreateAlembicCamera": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateArnoldAss": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ],
-            "ext": ".ass"
-        },
-        "CreateArnoldRop": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateCompositeSequence": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateHDA": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateKarmaROP": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateMantraIFD": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateMantraROP": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreatePointCache": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateBGEO": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateRedshiftProxy": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateRedshiftROP": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateReview": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateStaticMesh": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ],
-            "static_mesh_prefix": "S",
-            "collision_prefixes": [
-                "UBX",
-                "UCP",
-                "USP",
-                "UCX"
-            ]
-        },
-        "CreateUSD": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateUSDRender": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateVDBCache": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "CreateVrayROP": {
-            "enabled": true,
-            "default_variants": [
-                "Main"
-            ]
-        }
-    },
-    "publish": {
-        "CollectAssetHandles": {
-            "use_asset_handles": true
-        },
-        "CollectChunkSize": {
-            "enabled": true,
-            "optional": true,
-            "chunk_size": 999999
-        },
-        "ValidateContainers": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateMeshIsStatic": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateReviewColorspace": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateSubsetName": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateUnrealStaticMeshName": {
-            "enabled": false,
-            "optional": true,
-            "active": true
-        },
-        "ValidateWorkfilePaths": {
-            "enabled": true,
-            "optional": true,
-            "node_types": [
-                "file",
-                "alembic"
-            ],
-            "prohibited_vars": [
-                "$HIP",
-                "$JOB"
-            ]
-        }
-    }
-}
