--- client/ayon_core/settings/defaults/project_settings/kitsu.json
@@ -1,21 +1,0 @@
-{
-    "entities_naming_pattern": {
-        "episode": "E##",
-        "sequence": "SQ##",
-        "shot": "SH##"
-    },
-    "publish": {
-        "IntegrateKitsuNote": {
-            "set_status_note": false,
-            "note_status_shortname": "wfa",
-            "status_change_conditions": {
-                "status_conditions": [],
-                "family_requirements": []
-            },
-            "custom_comment_template": {
-                "enabled": false,
-                "comment_template": "{comment}\n\n|  |  |\n|--|--|\n| version| `{version}` |\n| family | `{family}` |\n| name | `{name}` |"
-            }
-        }
-    }
-}
