--- client/ayon_core/settings/defaults/project_settings/max.json
@@ -1,100 +1,0 @@
-{
-    "unit_scale_settings": {
-        "enabled": true,
-        "scene_unit_scale": "Meters"
-    },
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "RenderSettings": {
-        "default_render_image_folder": "renders/3dsmax",
-        "aov_separator": "underscore",
-        "image_format": "exr",
-        "multipass": true
-    },
-    "CreateReview": {
-        "review_width": 1920,
-        "review_height": 1080,
-        "percentSize": 100.0,
-        "keep_images": false,
-        "image_format": "png",
-        "visual_style": "Realistic",
-        "viewport_preset": "Quality",
-        "anti_aliasing": "None",
-        "vp_texture": true
-    },
-    "PointCloud": {
-        "attribute": {
-            "Age": "age",
-            "Radius": "radius",
-            "Position": "position",
-            "Rotation": "rotation",
-            "Scale": "scale",
-            "Velocity": "velocity",
-            "Color": "color",
-            "TextureCoordinate": "texcoord",
-            "MaterialID": "matid",
-            "custFloats": "custFloats",
-            "custVecs": "custVecs"
-        }
-    },
-    "publish": {
-        "ValidateFrameRange": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateAttributes": {
-            "enabled": false,
-            "attributes": {}
-        },
-        "ValidateCameraAttributes": {
-            "enabled": true,
-            "optional": true,
-            "active": false,
-            "fov": 45.0,
-            "nearrange": 0.0,
-            "farrange": 1000.0,
-            "nearclip": 1.0,
-            "farclip": 1000.0
-        },
-        "ValidateLoadedPlugin": {
-            "enabled": false,
-            "optional": true,
-            "family_plugins_mapping": []
-        },
-        "ExtractModelObj": {
-            "enabled": true,
-            "optional": true,
-            "active": false
-        },
-        "ExtractModelFbx": {
-            "enabled": true,
-            "optional": true,
-            "active": false
-        },
-        "ExtractModelUSD": {
-            "enabled": true,
-            "optional": true,
-            "active": false
-        },
-        "ExtractModel": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ExtractMaxSceneRaw": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        }
-    }
-}
