--- client/ayon_core/settings/defaults/project_settings/photoshop.json
@@ -1,91 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "remapping": {
-            "rules": []
-        },
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "create": {
-        "ImageCreator": {
-            "enabled": true,
-            "active_on_create": true,
-            "mark_for_review": false,
-            "default_variants": [
-                "Main"
-            ]
-        },
-        "AutoImageCreator": {
-            "enabled": false,
-            "active_on_create": true,
-            "mark_for_review": false,
-            "default_variant": ""
-        },
-        "ReviewCreator": {
-            "enabled": true,
-            "active_on_create": true,
-            "default_variant": ""
-        },
-        "WorkfileCreator": {
-            "enabled": true,
-            "active_on_create": true,
-            "default_variant": "Main"
-        }
-    },
-    "publish": {
-        "CollectColorCodedInstances": {
-            "enabled": true,
-            "create_flatten_image": "no",
-            "flatten_subset_template": "",
-            "color_code_mapping": []
-        },
-        "CollectReview": {
-            "enabled": true
-        },
-        "CollectVersion": {
-            "enabled": false
-        },
-        "ValidateContainers": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateNaming": {
-            "invalid_chars": "[ \\\\/+\\*\\?\\(\\)\\[\\]\\{\\}:,;]",
-            "replace_char": "_"
-        },
-        "ExtractImage": {
-            "formats": [
-                "png",
-                "jpg"
-            ]
-        },
-        "ExtractReview": {
-            "make_image_sequence": false,
-            "max_downscale_size": 8192,
-            "jpg_options": {
-                "tags": [
-                    "review",
-                    "ftrackreview"
-                ]
-            },
-            "mov_options": {
-                "tags": [
-                    "review",
-                    "ftrackreview"
-                ]
-            }
-        }
-    },
-    "workfile_builder": {
-        "create_first_version": false,
-        "custom_templates": []
-    }
-}
