--- client/ayon_core/settings/defaults/project_settings/resolve.json
@@ -1,35 +1,0 @@
-{
-    "launch_openpype_menu_on_start": false,
-    "imageio": {
-        "activate_host_color_management": true,
-        "remapping": {
-            "rules": []
-        },
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "create": {
-        "CreateShotClip": {
-            "hierarchy": "{folder}/{sequence}",
-            "clipRename": true,
-            "clipName": "{track}{sequence}{shot}",
-            "countFrom": 10,
-            "countSteps": 10,
-            "folder": "shots",
-            "episode": "ep01",
-            "sequence": "sq01",
-            "track": "{_track_}",
-            "shot": "sh###",
-            "vSyncOn": false,
-            "workfileFrameStart": 1001,
-            "handleStart": 10,
-            "handleEnd": 10
-        }
-    }
-}
