--- client/ayon_core/settings/defaults/project_settings/royalrender.json
@@ -1,10 +1,0 @@
-{
-    "rr_paths": [
-        "default"
-    ],
-    "publish": {
-        "CollectSequencesFromJob": {
-            "review": true
-        }
-    }
-}
