--- client/ayon_core/settings/defaults/project_settings/slack.json
@@ -1,20 +1,0 @@
-{
-    "token": "",
-    "publish": {
-        "CollectSlackFamilies": {
-            "enabled": true,
-            "optional": true,
-            "profiles": [
-                {
-                    "families": [],
-                    "hosts": [],
-                    "task_types": [],
-                    "tasks": [],
-                    "subsets": [],
-                    "review_upload_limit": 50.0,
-                    "channel_messages": []
-                }
-            ]
-        }
-    }
-}
