--- client/ayon_core/settings/defaults/project_settings/substancepainter.json
@@ -1,14 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "shelves": {}
-}
