--- client/ayon_core/settings/defaults/project_settings/tvpaint.json
@@ -1,111 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "stop_timer_on_application_exit": false,
-    "create": {
-        "create_workfile": {
-            "enabled": true,
-            "default_variant": "Main",
-            "default_variants": []
-        },
-        "create_review": {
-            "enabled": true,
-            "active_on_create": true,
-            "default_variant": "Main",
-            "default_variants": []
-        },
-        "create_render_scene": {
-            "enabled": true,
-            "active_on_create": false,
-            "mark_for_review": true,
-            "default_pass_name": "beauty",
-            "default_variant": "Main",
-            "default_variants": []
-        },
-        "create_render_layer": {
-            "mark_for_review": false,
-            "default_pass_name": "beauty",
-            "default_variant": "Main",
-            "default_variants": []
-        },
-        "create_render_pass": {
-            "mark_for_review": false,
-            "default_variant": "Main",
-            "default_variants": []
-        },
-        "auto_detect_render": {
-            "enabled": false,
-            "allow_group_rename": true,
-            "group_name_template": "L{group_index}",
-            "group_idx_offset": 10,
-            "group_idx_padding": 3
-        }
-    },
-    "publish": {
-        "CollectRenderInstances": {
-            "ignore_render_pass_transparency": false
-        },
-        "ExtractSequence": {
-            "review_bg": [
-                255,
-                255,
-                255,
-                255
-            ]
-        },
-        "ValidateProjectSettings": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateMarks": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ValidateStartFrame": {
-            "enabled": false,
-            "optional": true,
-            "active": true
-        },
-        "ValidateAssetName": {
-            "enabled": true,
-            "optional": true,
-            "active": true
-        },
-        "ExtractConvertToEXR": {
-            "enabled": false,
-            "replace_pngs": true,
-            "exr_compression": "ZIP"
-        }
-    },
-    "load": {
-        "LoadImage": {
-            "defaults": {
-                "stretch": true,
-                "timestretch": true,
-                "preload": true
-            }
-        },
-        "ImportImage": {
-            "defaults": {
-                "stretch": true,
-                "timestretch": true,
-                "preload": true
-            }
-        }
-    },
-    "workfile_builder": {
-        "create_first_version": false,
-        "custom_templates": []
-    }
-}
