--- client/ayon_core/settings/defaults/project_settings/unreal.json
@@ -1,21 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "level_sequences_for_layouts": false,
-    "delete_unmatched_assets": false,
-    "render_config_path": "",
-    "preroll_frames": 0,
-    "render_format": "png",
-    "project_setup": {
-        "dev_mode": false
-    }
-}
