--- client/ayon_core/settings/defaults/project_settings/webpublisher.json
@@ -1,145 +1,0 @@
-{
-    "imageio": {
-        "activate_host_color_management": true,
-        "ocio_config": {
-            "override_global_config": false,
-            "filepath": []
-        },
-        "file_rules": {
-            "activate_host_rules": false,
-            "rules": {}
-        }
-    },
-    "timeout_profiles": [
-        {
-            "hosts": [
-                "photoshop"
-            ],
-            "task_types": [],
-            "timeout": 600
-        }
-    ],
-    "publish": {
-        "CollectPublishedFiles": {
-            "sync_next_version": false,
-            "task_type_to_family": {
-                "Animation": [
-                    {
-                        "is_sequence": false,
-                        "extensions": [
-                            "tvp"
-                        ],
-                        "families": [],
-                        "tags": [],
-                        "result_family": "workfile"
-                    },
-                    {
-                        "is_sequence": true,
-                        "extensions": [
-                            "png",
-                            "exr",
-                            "tiff",
-                            "tif"
-                        ],
-                        "families": [
-                            "review"
-                        ],
-                        "tags": [
-                            "review"
-                        ],
-                        "result_family": "render"
-                    }
-                ],
-                "Compositing": [
-                    {
-                        "is_sequence": false,
-                        "extensions": [
-                            "aep"
-                        ],
-                        "families": [],
-                        "tags": [],
-                        "result_family": "workfile"
-                    },
-                    {
-                        "is_sequence": true,
-                        "extensions": [
-                            "png",
-                            "exr",
-                            "tiff",
-                            "tif"
-                        ],
-                        "families": [
-                            "review"
-                        ],
-                        "tags": [
-                            "review"
-                        ],
-                        "result_family": "render"
-                    }
-                ],
-                "Layout": [
-                    {
-                        "is_sequence": false,
-                        "extensions": [
-                            "psd"
-                        ],
-                        "families": [],
-                        "tags": [],
-                        "result_family": "workfile"
-                    },
-                    {
-                        "is_sequence": false,
-                        "extensions": [
-                            "png",
-                            "jpg",
-                            "jpeg",
-                            "tiff",
-                            "tif"
-                        ],
-                        "families": [
-                            "review"
-                        ],
-                        "tags": [
-                            "review"
-                        ],
-                        "result_family": "image"
-                    }
-                ],
-                "default_task_type": [
-                    {
-                        "is_sequence": false,
-                        "extensions": [
-                            "tvp",
-                            "psd"
-                        ],
-                        "families": [],
-                        "tags": [],
-                        "result_family": "workfile"
-                    },
-                    {
-                        "is_sequence": true,
-                        "extensions": [
-                            "png",
-                            "exr",
-                            "tiff",
-                            "tif"
-                        ],
-                        "families": [
-                            "review"
-                        ],
-                        "tags": [
-                            "review"
-                        ],
-                        "result_family": "render"
-                    }
-                ],
-                "__dynamic_keys_labels__": {
-                    "default_task_type": "Default task type"
-                }
-            }
-        },
-        "CollectTVPaintInstances": {
-            "layer_name_regex": "(?P<layer>L[0-9]{3}_\\w+)_(?P<pass>.+)"
-        }
-    }
-}
